requirements.txt
asttokens==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
daal==2025.4.0
debugpy==1.8.13
decorator==5.2.1
executing==2.2.0
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
imageio==2.37.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.4
nvidia-nccl-cu12==2.26.2
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==26.4.0
scikit-image==0.25.2
scikit-learn==1.6.1
scikit-learn-intelex==2025.4.0
scipy==1.15.2
setuptools==75.8.0
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
tbb==2022.1.0
tcmlib==1.3.0
threadpoolctl==3.6.0
tifffile==2025.3.30
tornado==6.4.2
tqdm==4.67.1
tqdm_joblib==0.0.4
traitlets==5.14.3
typing_extensions==4.13.1
tzdata==2025.2
umap-learn==0.5.7
wcwidth==0.2.13
wheel==0.45.1
xgboost==3.0.0
